% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDomesticWaterUsage.R
\name{calculateDomesticWaterUsage}
\alias{calculateDomesticWaterUsage}
\title{Total Water Usage Calculator}
\usage{
calculateDomesticWaterUsage(numberOfPeople, usageType)
}
\arguments{
\item{numberOfPeople}{Amount of people in the household}

\item{usageType}{Drinking, Bathing, or Cooking}
}
\value{
The output is the total estiimated usage in Litres per day \code{totalUsage}
}
\description{
Total Water Usage Calculator
}
