% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEnergyEmissions.R
\name{calculateEnergyEmissions}
\alias{calculateEnergyEmissions}
\title{Calculate Carbon Emissions from Energy Consumption}
\usage{
calculateEnergyEmissions(energyConsumption, energyType)
}
\arguments{
\item{energyConsumption}{Amount of energy consumed in kWh.}

\item{energyType}{Type of energy (e.g., 'electricity', 'natural_gas', 'oil').}
}
\value{
Estimated carbon emissions in kilograms of CO2 equivalent.
}
\description{
This function estimates the carbon emissions associated with energy consumption, considering different energy sources.
}
\examples{
calculateEnergyEmissions(200, 'electricity')
calculateEnergyEmissions(150, 'natural_gas')
}
