% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateTransportEmissions.R
\name{calculateTransportEmissions}
\alias{calculateTransportEmissions}
\title{Transportation Carbon Emissions Calculator}
\usage{
calculateTransportEmissions(distance, transportMode, efficiency)
}
\arguments{
\item{distance}{Sample distance traveled}

\item{transportMode}{The mode of transportation used}

\item{efficiency}{}
}
\value{
The output of \code{emissions}
}
\description{
Transportation Carbon Emissions Calculator
}
