% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateVehicleFuelEfficiency.R
\name{calculateVehicleFuelEfficiency}
\alias{calculateVehicleFuelEfficiency}
\title{Calculate Fuel Efficiency of Vehicles}
\usage{
calculateVehicleFuelEfficiency(fuelConsumed, distanceTraveled)
}
\arguments{
\item{fuelConsumed}{Fuel consumed in liters.}

\item{distanceTraveled}{Distance traveled in kilometers.}
}
\value{
Fuel efficiency in kilometers per liter.
}
\description{
Evaluates the fuel efficiency of vehicles based on the amount of fuel consumed and the distance traveled.
}
\examples{
calculateVehicleFuelEfficiency(50, 500)
calculateVehicleFuelEfficiency(70, 350)
}
