% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterUsageSummary.R
\name{summary,WaterUsage-method}
\alias{summary,WaterUsage-method}
\title{Summarize and visualize water usage data using ggplot2}
\usage{
\S4method{summary}{WaterUsage}(object)
}
\arguments{
\item{object}{An object of class WaterUsage.}
}
\value{
A list containing summary statistics and a ggplot.
}
\description{
This function provides summary statistics and a ggplot for water usage data stored in an S4 object.
}
\examples{
usageData <- new("WaterUsage", usageType = c("Domestic", "Industrial", "Agricultural"),
                 volume = c(150, 300, 450),
                 date = as.Date(c('2021-01-01', '2021-01-02', '2021-01-03')))
summary <- waterUsageSummary(usageData)
print(summary$plot)
}
